# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Building Maintenance API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/maintenance-requests/{id}": {
      "get": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "Get a maintenance request by ID",
        "description": "Returns a maintenance request by its unique identifier.",
        "operationId": "getOne_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "Update a maintenance request",
        "description": "Updates an existing maintenance request.",
        "operationId": "update_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Maintenance request updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "Delete a maintenance request",
        "description": "Deletes a maintenance request by its ID.",
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/maintenance-requests": {
      "get": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "List all maintenance requests",
        "description": "Returns a list of all maintenance requests.",
        "operationId": "listAll_3",
        "responses": {
          "200": {
            "description": "List of maintenance requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceRequestDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "Create a new maintenance request",
        "description": "Creates a new maintenance request.",
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Maintenance request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/residents/{residentId}/requests/{requestId}": {
      "get": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Get a maintenance request by ID for a resident",
        "description": "Returns a maintenance request by its unique identifier for the specified resident.",
        "operationId": "get",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Cancel a maintenance request for a resident",
        "description": "Cancels a maintenance request for the specified resident.",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponseDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Update a maintenance request for a resident",
        "description": "Updates an existing maintenance request for the specified resident.",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Maintenance request update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Maintenance request updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/maintenance-requests/{id}/technician/{technicianId}": {
      "patch": {
        "tags": [
          "Maintenance Request Management"
        ],
        "summary": "Assign a technician to a maintenance request",
        "description": "Assigns a technician to a maintenance request by their IDs.",
        "operationId": "assign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "technicianId",
            "in": "path",
            "description": "UUID of the technician",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Technician assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request or technician not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResidentialComplexDto": {
        "type": "object",
        "description": "Summary information about a residential complex.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the residential complex."
          },
          "name": {
            "type": "string",
            "description": "Name of the residential complex."
          },
          "address": {
            "type": "string",
            "description": "Address of the residential complex."
          },
          "city": {
            "type": "string",
            "description": "City where the residential complex is located."
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code of the residential complex."
          }
        }
      },
...
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "description",
          "scheduledAt"
        ]
      },
      "MaintenanceRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "resident": {
            "$ref": "#/components/schemas/ResidentSummaryDto"
          },
          "description": {
            "type": "string"
          },
          "specialization": {
            "type": "string",
            "enum": [
              "ELECTRICAL",
              "PLUMBING",
              "HVAC",
              "CARPENTRY",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "technician": {
            "$ref": "#/components/schemas/TechnicianSummaryDto"
          }
        }
      },
      "ResidentSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string"
          },
          "unitBlock": {
            "type": "string"
          },
          "residentialComplex": {
            "$ref": "#/components/schemas/ResidentialComplexDto"
          }
        }
      },
      "TechnicianSummaryDto": {
        "type": "object",
        "description": "Summary information about a technician.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the technician."
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the technician."
          },
          "email": {
            "type": "string",
            "description": "Email address of the technician."
          }
        }
      },
...
      "CreateMaintenanceRequestDto": {
        "type": "object",
        "description": "Payload for creating a new maintenance request. Resident ID is required.",
        "properties": {
          "residentId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the resident making the request",
            "example": "b1a7e8c2-1234-4f8a-9b2e-123456789abc"
          },
          "description": {
            "type": "string",
            "description": "Description of the maintenance issue",
            "example": "Leaking kitchen faucet",
            "minLength": 1
          },
          "specialization": {
            "type": "string",
            "description": "Required technician specialization",
            "enum": [
              "ELECTRICAL",
              "PLUMBING",
              "HVAC",
              "CARPENTRY",
              "OTHER"
            ],
            "example": "PLUMBING"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "description": "Scheduled date and time for the maintenance",
            "example": "2025-08-01T10:00:00"
          }
        },
        "required": [
          "description",
          "residentId",
          "scheduledAt",
          "specialization"
        ]
      },
...
      "DeleteResponseDto": {
        "type": "object",
        "description": "DTO returned after deleting a resource, containing a message about the operation result.",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the result of the delete operation.",
            "example": "Residential complex deleted successfully."
          }
        }
      },
      "CancelResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}